Changed Doubles to ints

+++ src/main/pl/edu/agh/to/core/CoreSystem.scala
@@ -80,7 +80,7 @@ object CoreSystem {
   private val testOperator = new Operator()
 
   private def testAgentProvider(operator: Operator): Agent = {
-    val genotype = new Genotype((Random.nextDouble()).toString)
+    val genotype = new Genotype((Random.nextInt()).toString)
     new Agent(genotype, 100, operator)
   }
